refactor: rename home and main feat
