+++ wear/src/main/AndroidManifest.xml
@@ -38,7 +38,7 @@
         <service
             android:name=".tile.MainTileService"
             android:exported="true"
-            android:label="@string/tile_label"
+            android:label="@string/app_name"
             android:permission="com.google.android.wearable.permission.BIND_TILE_PROVIDER">
             <intent-filter>
                 <action android:name="androidx.wear.tiles.action.BIND_TILE_PROVIDER" />
