+++ wear/src/main/java/fr/croumy/bouge/presentation/navigation/NavRoutes.kt
@@ -3,8 +3,8 @@ package fr.croumy.bouge.presentation.navigation
 open class NavRoutes(val route: String) {
     object Start: NavRoutes("Start")
     object PickCompanion: NavRoutes("PickCompanion")
-    object Home: NavRoutes("Home")
     object Main: NavRoutes("Main")
-    object Redirect: NavRoutes("Redirect")
+    object Home: NavRoutes("Home")
+    object Menu: NavRoutes("Menu")
     object ExercisesHistory: NavRoutes("ExercisesHistory")
 }
