+++ wear/src/main/java/fr/croumy/bouge/presentation/ui/screens/StartScreen.kt
@@ -45,7 +45,7 @@ fun StartScreen(
             startViewModel.initHealthService()
 
             navController.navigate(
-                if(hasCompanion) NavRoutes.Home.route else NavRoutes.PickCompanion.route
+                if(hasCompanion) NavRoutes.Main.route else NavRoutes.PickCompanion.route
             )
         }
     }
