+++ wear/src/main/java/fr/croumy/bouge/presentation/ui/screens/menu/MenuScreen.kt
@@ -14,7 +14,7 @@ import fr.croumy.bouge.presentation.ui.components.Button
 
 @OptIn(ExperimentalHorologistApi::class)
 @Composable
-fun RedirectScreen() {
+fun MenuScreen() {
     val navController = LocalNavController.current
 
     ScalingLazyColumn(
