+++ wear/src/main/java/fr/croumy/bouge/presentation/ui/screens/pickCompanion/PickCompanionScreen.kt
@@ -124,7 +124,7 @@ fun PickCompanionScreen(
             icon = Icons.Default.Check,
             onClick = {
                 pickCompanionViewModel.selectCompanion(selectedCompanion.value, customName.value)
-                navController.navigate(NavRoutes.Home.route)
+                navController.navigate(NavRoutes.Main.route)
             },
         )
     }
